+++ src/wifi/model/interference-helper.h
@@ -273,7 +273,7 @@ private:
    *
    * \return SNR in liear ratio
    */
-  double CalculateSnr (double signal, double noiseInterference, uint8_t channelWidth) const;
+  double CalculateSnr (double signal, double noiseInterference, uint16_t channelWidth) const;
   /**
    * Calculate the success rate of the chunk given the SINR, duration, and Wi-Fi mode.
    * The duration and mode are used to calculate how many bits are present in the chunk.
